+++ ChangeLog.rst
@@ -45,7 +45,7 @@ Added
 - New ``Catalogue`` class.
 - New ``Fingerprint`` class.
 - New ``canon_hash()`` function.
-- New geometry functions: ``centroid()``, `` rmsd()``, `` grmsd(), for_equiv_perms()``.
+- New geometry functions: ``centroid()``, ``rmsd()``, ``grmsd()``, ``for_equiv_perms()``.
 - New ``ortho_onto()`` function.
 - New ``Colour`` property.
 - New ``Geometry`` class.